Project_012/services/ui/requirements.txt
streamlit
requests
dotenv
pyyaml